Simplify assistant history helpers and tests
