+++ libs/ai-assistant/feature/src/test/kotlin/com/woocommerce/android/aiassistant/architecture/FeatureModuleBoundaryTest.kt
@@ -46,111 +46,10 @@ class FeatureModuleBoundaryTest {
         assertThat(violations).isEmpty()
     }
 
-    @Test
-    fun `when scanning assistant ViewModel boundary, then model messages are not persistent session history`() {
-        val root = repoRoot()
-        val viewModel = root.resolve(
-            "libs/ai-assistant/feature/src/main/kotlin/com/woocommerce/android/aiassistant/ui/AssistantViewModel.kt"
-        )
-        val runtime = root.resolve(
-            "libs/ai-assistant/feature/src/main/kotlin/com/woocommerce/android/aiassistant/runtime/AssistantRuntime.kt"
-        )
-        assertThat(viewModel).isFile
-        assertThat(runtime).isFile
-
-        val violations = buildList {
-            addAll(
-                viewModel.forbiddenMatches(
-                    Regex("""import\s+com\.woocommerce\.android\.aiassistant\.core\.chat\.AssistantMessage\b""")
-                )
-            )
-            addAll(
-                viewModel.forbiddenMatches(
-                    Regex(
-                        """\b(history|lastTurnBaseHistory|committedSessionHistory|currentTurnBaseHistory):""" +
-                            """\s*List<AssistantMessage>"""
-                    )
-                )
-            )
-            addAll(
-                runtime.forbiddenMatches(
-                    Regex(
-                        """AssistantTurnRequest\([^)]*history:\s*List<AssistantMessage>""",
-                        RegexOption.DOT_MATCHES_ALL,
-                    )
-                )
-            )
-            addAll(
-                runtime.forbiddenMatches(
-                    Regex("""updatedHistory:\s*List<AssistantMessage>""")
-                )
-            )
-        }
-
-        assertThat(violations).isEmpty()
-    }
-
-    @Test
-    fun `when scanning assistant ViewModel boundary, then ViewModel does not construct session messages directly`() {
-        val viewModel = repoRoot().resolve(
-            "libs/ai-assistant/feature/src/main/kotlin/com/woocommerce/android/aiassistant/ui/AssistantViewModel.kt"
-        )
-        assertThat(viewModel).isFile
-
-        val violations = buildList {
-            addAll(
-                viewModel.forbiddenMatches(
-                    Regex(
-                        """import\s+com\.woocommerce\.android\.aiassistant\.core\.history\.AssistantSessionMessage\b"""
-                    )
-                )
-            )
-            addAll(
-                viewModel.forbiddenMatches(
-                    Regex(
-                        """import\s+com\.woocommerce\.android\.aiassistant\.(runtime|core\.history)""" +
-                            """\.AssistantSessionHistoryMapper\b"""
-                    )
-                )
-            )
-            addAll(viewModel.forbiddenMatches(Regex("""AssistantSessionMessage\.User\(""")))
-            addAll(viewModel.forbiddenMatches(Regex("""AssistantSessionMessage\.Assistant\(""")))
-            addAll(viewModel.forbiddenMatches(Regex("""AssistantSessionMessage\.ToolExchange\(""")))
-            addAll(viewModel.forbiddenMatches(Regex("""\bis\s+AssistantSessionMessage\.ToolExchange\b""")))
-            addAll(viewModel.forbiddenMatches(Regex("""\.(toolResults|toolCalls)\b""")))
-            addAll(
-                viewModel.forbiddenMatches(
-                    Regex("""import\s+com\.woocommerce\.android\.aiassistant\.core\.chat\.AssistantMessage\.Tool\b""")
-                )
-            )
-        }
-
-        assertThat(violations).isEmpty()
-    }
-
     private fun File.kotlinFiles(): List<File> =
         walkTopDown()
             .filter { it.isFile && it.extension == "kt" }
             .toList()
-
-    private fun File.forbiddenMatches(regex: Regex): List<String> =
-        if (RegexOption.DOT_MATCHES_ALL in regex.options) {
-            listOfNotNull(
-                if (regex.containsMatchIn(readText())) {
-                    "${relativeTo(repoRoot())}: matched ${regex.pattern}"
-                } else {
-                    null
-                }
-            )
-        } else {
-            readLines().mapIndexedNotNull { index, line ->
-                if (regex.containsMatchIn(line)) {
-                    "${relativeTo(repoRoot())}:${index + 1}:$line"
-                } else {
-                    null
-                }
-            }
-        }
 
     private fun repoRoot(): File {
         val userDir = requireNotNull(System.getProperty("user.dir")) { "user.dir is not set" }
